+++ src/const.js
@@ -1,0 +1,28 @@
+export const CardCount = {
+  DEFAULT: 21,
+  ON_START: 5,
+  STEP: 5,
+  EXTRA: 2
+};
+
+export const SortType = {
+  DEFAULT: `Sort by default`,
+  DATE: `Sort by date`,
+  RATING: `Sort by rating`,
+};
+
+export const FILM_GENRES = [
+  `Horror`,
+  `Action`,
+  `Adventure`,
+  `Musical`,
+  `Comedy`,
+  `Thriller`,
+  `Detective`,
+  `Anime`,
+  `Drama`,
+  `Melodrama`,
+  `Art house`,
+  `Fantasy`,
+  `Space opera`,
+];
